moving to REST way + datasource factories and services

+++ src/main/java/org/example/ataccama/routers/DatabaseConnectionRouter.java
@@ -5,28 +5,36 @@ import lombok.extern.slf4j.Slf4j;
 import org.example.ataccama.data.DatabaseConnection;
 import org.example.ataccama.repos.DatabaseConnectionRepo;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
 
 import static org.example.ataccama.data.DatabaseConnection.DATABASE_COLUMN;
 import static org.example.ataccama.data.DatabaseConnection.HOST_COLUMN;
 import static org.example.ataccama.data.DatabaseConnection.NAME_COLUMN;
+import static org.example.ataccama.routers.DatabaseConnectionRouter.CONNECTIONS_PATH;
 import static org.springframework.data.domain.Sort.Direction.ASC;
-import static org.springframework.http.HttpStatus.OK;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static org.springframework.http.ResponseEntity.badRequest;
+import static org.springframework.http.ResponseEntity.created;
+import static org.springframework.http.ResponseEntity.notFound;
+import static org.springframework.http.ResponseEntity.ok;
 
-@Controller
+@RestController
+@RequestMapping("/" + CONNECTIONS_PATH)
 @Slf4j
 @RequiredArgsConstructor
 public class DatabaseConnectionRouter {
-    private static final String CONNECTIONS_PATH = "connections";
-    private static final String CONNECTION_PATH = "connection";
+    public static final String CONNECTIONS_PATH = "connections";
+    public static final String ID_PATH_VAR = "id";
 
     private static final int DEFAULT_PAGE = 1;
     private static final int DEFAULT_MIN_PAGE_SIZE = 1;
@@ -34,50 +42,54 @@ public class DatabaseConnectionRouter {
 
     private final DatabaseConnectionRepo databaseConnectionRepo;
 
-    @GetMapping("/" + CONNECTIONS_PATH)
-    public String get(//
-                      @RequestParam(defaultValue = "" + DEFAULT_PAGE) int page, //
-                      @RequestParam(defaultValue = "" + DEFAULT_PAGE_SIZE) int pageSize, //
-
-                      Model model) {
-        model.addAttribute("connection", new DatabaseConnection());
-        model.addAttribute("connections", databaseConnectionRepo.findAll(PageRequest.of(Math.max(DEFAULT_PAGE, page) - 1, //
-                                                                                        Math.max(DEFAULT_MIN_PAGE_SIZE, pageSize), //
-                                                                                        ASC, //
-                                                                                        NAME_COLUMN, HOST_COLUMN, DATABASE_COLUMN))
-                                                                .getContent());
-        return CONNECTIONS_PATH;
+    @GetMapping(produces = APPLICATION_JSON_VALUE)
+    public ResponseEntity<?> readAll(@RequestParam(defaultValue = "" + DEFAULT_PAGE) int page, //
+                                     @RequestParam(defaultValue = "" + DEFAULT_PAGE_SIZE) int pageSize) {
+        return ok(databaseConnectionRepo.findAll(PageRequest.of(Math.max(DEFAULT_PAGE, page) - 1, //
+                                                                Math.max(DEFAULT_MIN_PAGE_SIZE, pageSize), //
+                                                                ASC, //
+                                                                NAME_COLUMN, HOST_COLUMN, DATABASE_COLUMN))
+                                        .getContent());
     }
 
-    @GetMapping("/" + CONNECTIONS_PATH + "/{id}")
-    public String getById(@PathVariable(name = "id") Long id, Model model) {
+    @GetMapping(value = "/{" + ID_PATH_VAR + "}", //
+            produces = APPLICATION_JSON_VALUE)
+    public ResponseEntity<?> readOne(@PathVariable Long id) {
         final var databaseConnection = databaseConnectionRepo.findById(id);
-        if (databaseConnection.isPresent()) {
-            model.addAttribute("connection", databaseConnection.get());
-            return CONNECTION_PATH;
-        }
-        return "redirect:/" + CONNECTIONS_PATH;
+        return databaseConnection.map(ResponseEntity::ok)
+                                 .orElseGet(notFound()::build);
     }
 
-    @PostMapping("/" + CONNECTIONS_PATH)
-    public String post(@ModelAttribute DatabaseConnection connection) {
+    @PostMapping(consumes = APPLICATION_JSON_VALUE)
+    public ResponseEntity<?> create(@RequestBody DatabaseConnection connection) {
+        if (connection != null) {
+            final var databaseConnection = databaseConnectionRepo.save(connection);
+            return created(URI.create("/" + CONNECTIONS_PATH + "/" + databaseConnection.getId())).build();
+        }
+        return badRequest().build();
+    }
+
+    @PostMapping(value = "/{" + ID_PATH_VAR + "}", //
+            consumes = APPLICATION_JSON_VALUE)
+    public ResponseEntity<?> update(@PathVariable Long id, //
+                                    @RequestBody DatabaseConnection connection) {
+        final var databaseConnection = databaseConnectionRepo.findById(id);
+        if (databaseConnection.isEmpty()) {
+            return notFound().build();
+        }
+        if (connection == null) {
+            return badRequest().build();
+        }
         databaseConnectionRepo.save(connection);
-        return "redirect:/" + CONNECTIONS_PATH;
+        return ok().build();
     }
 
-    @PostMapping("/" + CONNECTIONS_PATH + "/{id}")
-    public String postById(@PathVariable(name = "id") Long id, @ModelAttribute DatabaseConnection connection, Model model) {
-        final var databaseConnection = databaseConnectionRepo.findById(id);
-        if (databaseConnection.isPresent()) {
-            model.addAttribute("connection", databaseConnectionRepo.save(connection));
-            return CONNECTION_PATH;
+    @DeleteMapping("/{" + ID_PATH_VAR + "}")
+    public ResponseEntity<?> delete(@PathVariable Long id) {
+        if (databaseConnectionRepo.existsById(id)) {
+            databaseConnectionRepo.deleteById(id);
+            return ok().build();
         }
-        return "redirect:/" + CONNECTIONS_PATH;
-    }
-
-    @DeleteMapping("/" + CONNECTIONS_PATH + "/{id}")
-    @ResponseStatus(value = OK)
-    public void delete(@PathVariable Long id) {
-        databaseConnectionRepo.deleteById(id);
+        return notFound().build();
     }
 }
